Fix circular imports - Lol, very short lived lessonIndices.ts since it's getting axed in the next PR
