+++ website/src/utils/timetables/index.ts
@@ -1,22 +1,5 @@
 import { AcadWeekInfo } from 'nusmoderator';
-import {
-  entries,
-  filter,
-  first,
-  flatMapDeep,
-  get,
-  groupBy,
-  intersection,
-  isEqual,
-  map,
-  mapValues,
-  maxBy,
-  pick,
-  range,
-  reduce,
-  sample,
-  values,
-} from 'lodash-es';
+import { flatMapDeep, groupBy, isEqual, map, mapValues, range, sample, values } from 'lodash-es';
 import { addDays, min as minDate, parseISO, startOfDay } from 'date-fns';
 
 import {
@@ -24,12 +7,10 @@ import {
   LessonIndex,
   LessonType,
   RawLessonWithIndex,
-  Module,
   NumericWeeks,
   RawLesson,
   Semester,
   ClassNo,
-  LessonIndicesMap,
 } from 'types/modules';
 
 import {
@@ -41,8 +22,6 @@ import {
   SemTimetableConfig,
   SemTimetableConfigWithLessons,
 } from 'types/timetables';
-
-import { ModulesMap } from 'types/reducers';
 
 import { getTimeAsDate } from '../timify';
 import { deltas } from '../array';
@@ -130,65 +109,6 @@ export function getStartTimeAsDate(lesson: Lesson, date: Date = new Date()): Dat
 
 export function getEndTimeAsDate(lesson: Lesson, date: Date = new Date()): Date {
   return getTimeAsDate(lesson.endTime, date);
-}
-
-/**
- * Used to recover from the config of a lesson type that contains invalid lesson indices
- * @param lessonsWithLessonType lessons with the same lesson type to generate a valid lesson config from
- * @returns lesson indices of the generated valid lesson config
- *
- * Note: the current implementation generates a config containing lessons belonging to the first classNo in the provided lessons
- */
-export function getRecoveryLessonIndices(
-  lessonsWithLessonType: RawLessonWithIndex[],
-): LessonIndex[] {
-  const firstClass = first(lessonsWithLessonType);
-  if (!firstClass) {
-    return [];
-  }
-  const { classNo } = firstClass;
-  const validLessonIndices = map(
-    filter(lessonsWithLessonType, (lesson) => lesson.classNo === classNo),
-    'lessonIndex',
-  );
-  return validLessonIndices;
-}
-
-/**
- * Group lessons by lesson types then classNo
- * @param lessonsWithIndex lessons to group
- * @returns lesson indices, not lessons
- */
-export const makeLessonIndicesMap = (
-  lessonsWithIndex: readonly RawLessonWithIndex[],
-): LessonIndicesMap => {
-  const lessonsByLessonType = groupBy(lessonsWithIndex, 'lessonType');
-  return mapValues(lessonsByLessonType, (lessonsWithLessonType) => {
-    const lessonsByClassNo = groupBy(lessonsWithLessonType, 'classNo');
-    return mapValues(lessonsByClassNo, (lessonsWithClassNo) =>
-      map(lessonsWithClassNo, 'lessonIndex'),
-    );
-  });
-};
-
-/**
- * Helper function to return the indices of lessons belonging to the {@link LessonType|lesson type} and {@link ClassNo|classNo} in the {@link LessonIndicesMap|lesson index mapping}
- * @param lessonIndicesMap
- * @param lessonType
- * @param classNo
- */
-export const getLessonIndices = (
-  lessonIndicesMap: LessonIndicesMap,
-  lessonType: LessonType,
-  classNo: ClassNo,
-): LessonIndex[] => get(get(lessonIndicesMap, lessonType), classNo);
-
-// Get information for all modules present in a semester timetable config
-export function getSemesterModules(
-  timetable: { [moduleCode: string]: unknown },
-  modules: ModulesMap,
-): Module[] {
-  return values(pick(modules, Object.keys(timetable)));
 }
 
 /**
@@ -288,52 +208,16 @@ export function getHoverLesson(lesson: InteractableLesson): HoverLesson {
   };
 }
 
-/**
- * Based on what lessons are currently in the lesson config, find the classNo that most of the lessons belong to
- * @param lessonIndicesMap {@link LessonIndicesMap|Lesson indices mapping} of the module
- * @param timetableLessonIndices lessons currently in lesson config
- * @returns a lesson config consisting of lesson indices that best matches the TA lesson config
- */
-export function getClosestLessonConfig(
-  lessonIndicesMap: LessonIndicesMap,
-  timetableLessonIndices: ModuleLessonConfig,
-): ModuleLessonConfig {
-  return reduce(
-    lessonIndicesMap,
-    (accumulatedModuleLessonConfig, lessonsWithLessonType, lessonType) => {
-      const timetableLessonsWithLessonType = timetableLessonIndices[lessonType];
-      const lessonGroupOccurrences = entries(
-        reduce(
-          lessonsWithLessonType,
-          (accumulated, lessonIndices, lessonGroup) => ({
-            ...accumulated,
-            [lessonGroup]: intersection(lessonIndices, timetableLessonsWithLessonType).length,
-          }),
-          {} as Record<ClassNo, number>,
-        ),
-      );
-
-      const closestLessonGroups = maxBy(lessonGroupOccurrences, ([, occurrences]) => occurrences);
-      if (!closestLessonGroups) return accumulatedModuleLessonConfig;
-      const [closestLessonGroupKey] = closestLessonGroups;
-      const closestLessonGroup = getLessonIndices(
-        lessonIndicesMap,
-        lessonType,
-        closestLessonGroupKey,
-      );
-
-      return {
-        ...accumulatedModuleLessonConfig,
-        [lessonType]: closestLessonGroup,
-      };
-    },
-    {} as ModuleLessonConfig,
-  );
-}
-
 export { findExamClashes } from './exams';
 export { isInteractable, getInteractableLessons } from './interactabilityHydration';
 export { hydrateSemTimetableWithLessons } from './lessonHydration';
+export {
+  getClosestLessonConfig,
+  getLessonIndices,
+  getRecoveryLessonIndices,
+  getSemesterModules,
+  makeLessonIndicesMap,
+} from './lessonIndices';
 export { LESSON_ABBREV_TYPE, LESSON_TYPE_ABBREV, getLessonIdentifier } from './lessonId';
 export { arrangeLessonsForWeek, groupLessonsByDay } from './lessonsArrangement';
 export { migrateSemTimetableConfig } from './migration';
